requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.1
pymongo==3.9.0
six==1.13.0
soupsieve==1.9.5
Werkzeug==0.16.0
mysql-connector-python~=8.0.18
requests~=2.27.1
selenium~=3.141.0
chardet==3.0.4
idna==2.10
urllib3==1.25.10